% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genre_list.R
\name{get_genre_list}
\alias{get_genre_list}
\title{GET request to RAWG}
\usage{
get_genre_list(api_key = "")
}
\arguments{
\item{api_key}{(str): your api key (recommended - not required, default = none)}
}
\value{
Return Data.Frame with list of genres
}
\description{
Submit RAWG query and receive RAWG response for game genres
}
\examples{
test<-get_genre_list()
}
