% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_platform.R
\name{get_platform}
\alias{get_platform}
\title{GET request to RAWG}
\usage{
get_platform(api_key = "", ordering = "added", Page = "", Page_size = "")
}
\arguments{
\item{api_key}{(str): your api key (recommended - not required, default = none)}

\item{ordering}{(str): Which field to use when ordering the results("id", "added")}

\item{Page}{(int): A page number within the paginated result set.}

\item{Page_size}{(int): Number of results to return per page.}
}
\value{
Return Data.Frame with list of game platforms
}
\description{
Submit RAWG query and receive RAWG response formatted as a data frame
}
\examples{
key = "95731a37dcf84a638308f6f40fb01801"
p = 1
p_size = 5
ordering = "id"
df <- get_platform(key,Page = p, Page_size = p_size, ordering)
}
