% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publisher_list.R
\name{get_publisher_list}
\alias{get_publisher_list}
\title{GET request to RAWG}
\usage{
get_publisher_list(api_key = "")
}
\arguments{
\item{api_key}{(str): your api key (recommended - not required, default = none)}
}
\value{
Return Data.Frame with list of publishers
}
\description{
Submit RAWG query and receive RAWG response for game publishers
}
\examples{
test<-get_publisher_list()
}
